+++ server/game/iron_golem.h
@@ -6,10 +6,9 @@
 
 #include "enemy_npc.h"
 
-// Game's hardest enemy
 class IronGolem: public EnemyNpc {
 public:
-    IronGolem(Position position, Rng& rng, EquipableItems& equipable_items, uint32_t level);
+    IronGolem(Position position, Rng& rng, EquipableItems& equipable_items, uint8_t level);
 };
 
 #endif
