+++ server/game/orc.h
@@ -6,10 +6,9 @@
 
 #include "enemy_npc.h"
 
-// Open world's strong enemy
 class Orc: public EnemyNpc {
 public:
-    Orc(Position position, Rng& rng, EquipableItems& equipable_items, uint32_t level);
+    Orc(Position position, Rng& rng, EquipableItems& equipable_items, uint8_t level);
 };
 
 #endif
